Version demo

+++ database/migrations/2020_02_11_222321_create_threads_table.php
@@ -18,7 +18,7 @@ class CreateThreadsTable extends Migration
 
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('channel_id');
-            $table->unsignedBigInteger('replies_count')->default(0);
+            $table->integer('replies_count')->default(0);
             $table->unsignedBigInteger('visits')->default(0);
             $table->string('title');
             $table->bigInteger('best_reply_id')->nullable();
